Cleaned main analysis test (3 MS+GC iterations)
